Handle case where page is reloaded without Vue
If the page is reloaded during wait and if the new content does not include Vue, the script used in setup_ready will fail because Vue is no longer defined. A reload like this can happen if a test case clicks e.g. a download-as-pdf  link. We need to guard against this edge case and set window.vueReady to true when it occurs, without attempting to wait for nextTick.